missing ~/.fluxable now auto-added&msgs improved
There was a potential problem when installed with pacman on arch if the user did not manually create her ~/.fluxable directory. Now this is checked for
each run of the main script and the dir is added if missing.
Messging is is improvedwhen run from outside of a graphical session.
Minor code realignment and comment updates...